+++ components/portal/messages/services.js
@@ -106,7 +106,8 @@ define(['angular'], function(angular) {
                 var added = false;
                 // If the message's groups array has groups,
                 // check for matches against portal groups
-                if (message.audienceFilter.groups.length > 0) {
+                if (message.audienceFilter && message.audienceFilter.groups
+                  && message.audienceFilter.groups.length > 0) {
                   // For each group for the current message
                   angular.forEach(message.audienceFilter.groups,
                     function(messageGroup) {
@@ -123,8 +124,8 @@ define(['angular'], function(angular) {
                       }
                     });
                 } else {
-                  // If the message's groups array is empty or null,
-                  // show it to everyone
+                  // If the message's groups array is empty or null or
+                  // non-existent, show it to everyone
                   messagesByGroup.push(message);
                   added = true;
                 }